#2ad4gkt[in progress] Add ProofOfStakeState and CommitteeSelector
ProofOfStakeState allow validators nodes to keep in memory the stake
of all stakers.

CommitteeSelector use ProofOfStakeState to compute randomly a committee.
SimpleCommitteeSelector is a CommitteeSelector which compute randomly
committee however the probability to be in the committee is proportional
to weight of the staker. More the weight is great, more the chance to be
in the committee is high.